requirements.txt
Flask==3.0.3
Pillow==10.4.0
rembg==2.0.59
